+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-### Fixed:
-* Potentially reduced client load.
-* Leveling data client crash when selecting SoA choice under certain circumstances in a server.
+### Added:
+* Added **/kingdomkeys check** command for server owners to see other player stats and abilities.